==> Step_1_Data_Cleaning_Extraction.ipynb <==
# importing the dependencies
import pandas as pd
import numpy
import json
from json import loads, dumps

##Cancer Rates Data##

# Making the dataframe to inspect the data
df_cancer = pd.read_csv("Source_Data/Countytable.csv")
df_cancer.head()

# getting rid of some columns and renaming some columns
df_cancer.drop(["Cancer header", "Note","Cancer type","Sex"], axis = 1,inplace = True)
df_cancer.head()

df_cancer = df_cancer.rename(columns = {"Rate": "cancer_rate_per_100k","County population": "county_population","County":"county"})
df_cancer.head()

# sort by county name 
df_cancer.sort_values(by = "county",inplace = True)
df_cancer.head()

# setting the index
df_cancer.reset_index(drop= True, inplace = True)
df_cancer.head()

# checking for null values
df_cancer.isnull().sum()

# checking the duplicated values
df_cancer.duplicated()

#checking the data types
df_cancer.dtypes

# change the datatype of county population to integer
df_cancer['county_population'] = df_cancer['county_population'].str.replace(',', '').astype(float)
#df_cancer[['County population']].apply(pd.to_numeric) 
#df_cancer.dtypes
df_cancer.head()

df_cancer.dtypes

## Health Outcomes Data ##

#reading the healthoutcome excel file 
df_healthoutcome = pd.read_excel("Source_Data/HealthOutcomeAndFactors.xlsx",header = 1)
df_healthoutcome.head()

#Renaming the columns
df_healthoutcome.rename( columns = {"Z-Score":"hlt_outcome_z", 
                                    "Rank":"hlt_outcome_rank","Z-Score.1":"hlt_factor_z",
                                    "Rank.1":"hlt_factor_rank","County": "county"}, inplace = True)
df_healthoutcome.head()

df_healthoutcome.drop(index = 0,inplace = True)
df_healthoutcome.head()

# resetting the index
df_healthoutcome.reset_index( drop = True, inplace = True)
df_healthoutcome.head()

# dropping column 
df_healthoutcome.drop( columns = ["State"], inplace = True)
df_healthoutcome.head()

## Asthma Data ##

#reading the asthma data
df_asthma = pd.read_csv("Source_Data/MN-asthma-county.csv")
df_asthma.head()

# Getting the column names
df_asthma.columns

 # dropping some columns
df_asthma.drop(['__Outcome', '__Year','Note'], axis = 1,inplace = True)
df_asthma.head()

# Renaming the columns
df_asthma.rename(columns = {"_County": "county","Age-adjusted rate per 10,000":"asthma_rate_per_100k","Count of cases":
                           "count_of_cases"}, inplace = True)
df_asthma.head()

# sorting data by county names
df_asthma.sort_values(by = "county", inplace = True)
df_asthma.head()

# Resetting the index
df_asthma.reset_index(drop= True, inplace = True)
df_asthma

# checking the length of dataframe
len(df_asthma)

#Checking for null values
df_asthma.isnull().sum()

# checking for data types 
df_asthma.dtypes

df_asthma['asthma_rate_per_100k'] = pd.to_numeric(df_asthma['asthma_rate_per_100k']
                                                                           ,errors='coerce')
df_asthma.dtypes

df_asthma['count_of_cases'] = pd.to_numeric(df_asthma['count_of_cases'], errors='coerce')
                                                                           
df_asthma.dtypes

df_asthma.head()

## EPA FRS Sites Data ##

# Read in CSV file to create EPA FRS dataframe
df_epa_frs = pd.read_csv("Source_Data/STATE_SINGLE_MN.csv")

# inspect dataframe 
df_epa_frs.head()

# inspect columns
df_epa_frs.columns

# create new dataframe by dropping unneccessary columns
df_epa_frs_clean = df_epa_frs[['REGISTRY_ID', 'PRIMARY_NAME', 'LOCATION_ADDRESS', 'COUNTY_NAME','SITE_TYPE_NAME', 'INTEREST_TYPES', 'LATITUDE83', 'LONGITUDE83']]

# rename columns for readability
df_epa_frs_clean.rename(columns = {'LATITUDE83': 'LATITUDE', 'LONGITUDE83': 'LONGITUDE', 'PRIMARY_NAME': 'SITE PRIMARY NAME', 'LOCATION_ADDRESS': 'SITE_ADDRESS'}, inplace = True)

#drop rows with null values in coordinate, interest types, location address, or site types columns
df_epa_frs_clean = df_epa_frs_clean.dropna(axis=0, subset=['LATITUDE'])
df_epa_frs_clean = df_epa_frs_clean.dropna(axis=0, subset=['LONGITUDE'])

# drop rows where interest type is "state master" only
df_epa_frs_clean = df_epa_frs_clean[df_epa_frs_clean.INTEREST_TYPES != "STATE MASTER"]

# create list of search terms for interest types to try and narrow down to relevant sites
search_term_list = ['RELEASE ASSESSMENT','AIR EMISSION INVENTORY', 'AIR MAJOR', 'AIR MINOR', 'AIR SYNTHETIC MINOR', 'ELECTRIC GENERATOR', 'ELECTRIC POWER GENERAT', 'ETHANOL FACILITY', 'GASOLINE', 'GREENHOUSE GAS', 'ANIMAL', 'LIVESTOCK', 'TRI REPORTER', 'TSCA SUBMITTER', 'EPCRA', 'FRP', 'SPCC', 'MINERAL EXPLORATION', 'CATHODE RAY', 'CESQG', 'CORRECTIVE ACTION', 'HAZARDOUS WASTE', 'INCINERATOR', 'INFECTIOUS WASTE', 'LQG', 'SQG', 'TSD', 'UIC', 'USED OIL', 'COMPLIANCE', 'ENFORCEMENT', 'PESTICIDE', 'RAD NPL', 'RAD WIPP', 'RAD NESHAPS', 'RADIOACTIVE', 'CLEANUP', 'REMEDIATION', 'REMEDIAL', 'BROWNFIELDS', 'SEDIMENTS', 'SUPERFUND', 'E-WASTE', 'INDUSTRIAL SITES', 'MATERIAL RECYCLING', 'MINE', 'MINING', 'OIL CONTROL', 'REFUSE DISPOSAL', 'SCRAP TIRE MANAGEMENT', 'SEWAGE SLUDGE', 'SOLID WASTE', 'LEAKING STORAGE TANK','UNDERGROUND STORAGE TANK', 'BIOSOLIDS', 'NPDES MAJOR', 'NPDES-MINOR', 'NPDES NON-MAJOR' 'NPDES UNPERMITTED', 'OIL AND GAS', 'MULTI-PERMITTED SITE']

# filter dataframe by search term list within the Interest_Types field
final_epa_frs_clean = df_epa_frs_clean[df_epa_frs_clean['INTEREST_TYPES'].str.contains('|'.join(search_term_list), na=False)]

# modified list of search terms for a smaller, more precise dataset focused on pollution
modified_search_term_list = ['TRI REPORTER', 'BROWNFIELDS PROPERTY', 'PESTICIDE PRODUCER', 'LQG', 'AIR MAJOR', 'TSCA SUBMITTER', 'HAZARDOUS AIR POLLUTANT MAJOR', 'GREENHOUSE GAS REPORTER', 'SUPERFUND', 'GASOLINE AND DIESEL PRODUCERS', 'ICIS-NPDES MAJOR', 'LEAKING UNDERGROUND STORAGE TANK', 'OIL BASED', 'GAS BASED', 'TSD', 'COAL BASED']

# filter dataframe by search term list within the Interest_Types field
modified_final_epa_frs_clean = df_epa_frs_clean[df_epa_frs_clean['INTEREST_TYPES'].str.contains('|'.join(modified_search_term_list), na=False)]

# filter out certain strings
modified_final_epa_frs_clean[modified_final_epa_frs_clean["INTEREST_TYPES"].str.contains('VSQG')==False]
modified_final_epa_frs_clean[modified_final_epa_frs_clean["INTEREST_TYPES"].str.contains('SQG')==False]

# inspect final dataframe
final_epa_frs_clean.tail(10)

modified_final_epa_frs_clean.tail(10)

## JSONIFY dataframes for ingestion into DB ##

# saving the data in an output file as json format
df_cancer.to_json('Cleaned_Data/cancer(2015-2019).json', orient='records')

# Saving the file in json format
df_healthoutcome.to_json('Cleaned_Data/healthoutcome(2023).json', orient='records')

# saving the file in json form
df_asthma.to_json('Cleaned_Data/asthma(2018-2020).json', orient='records')

# convert EPA DFs to json
final_epa_frs_clean.to_json('Cleaned_Data/epa_frs_data.json', orient='records')
modified_final_epa_frs_clean.to_json('Cleaned_Data/modified_final_epa_frs_clean.json', orient='records')
